Build  chartVM
